main project

+++ frontend/src/api/AppointmentApi.js
@@ -2,39 +2,54 @@
 const API_URL = "https://hospital-patient-care-dashboard-backend.onrender.com/api/appointment";
 
 export const BookAppointment = async (data) => {
-    try {
-        const res = await fetch(`${API_URL}/create`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify(data),
-        });
-        return await res.json();
-    } catch (err) {
-        console.error("BookAppointment error:", err);
-        return { success: false, message: "Network error" };
-    }
+  try {
+    const token = localStorage.getItem("token"); //  JWT token
+    if (!token) return { success: false, message: "User not logged in" };
+    const res = await fetch(`${API_URL}/create`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+        'Authorization': `Bearer ${token}`
+      },
+      body: JSON.stringify(data),
+    });
+    return await res.json();
+  } catch (err) {
+    console.error("BookAppointment error:", err);
+    return { success: false, message: "Network error" };
+  }
 };
 
 
 export const GetAllAppointment = async () => {
-    try {
-        const res = await fetch(`${API_URL}/`, {
-            method: "GET",
-            headers: { "Content-Type": "application/json" },
-           
-        });
-        return await res.json();
-    } catch (err) {
-        console.error("BookAppointment error:", err);
-        return { success: false, message: "Network error" };
-    }
+  try {
+    const token = localStorage.getItem("token"); //  JWT token
+    if (!token) return { success: false, message: "User not logged in" };
+    const res = await fetch(`${API_URL}/`, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+        'Authorization': `Bearer ${token}`
+      },
+
+    });
+    return await res.json();
+  } catch (err) {
+    console.error("BookAppointment error:", err);
+    return { success: false, message: "Network error" };
+  }
 };
 // discharge
 export const CompleteAppointment = async (id) => {
   try {
+    const token = localStorage.getItem("token"); //  JWT token
+    if (!token) return { success: false, message: "User not logged in" };
     const res = await fetch(`${API_URL}/update/${id}`, {
       method: "PUT",
-      headers: { "Content-Type": "application/json" },
+      headers: {
+        "Content-Type": "application/json",
+        'Authorization': `Bearer ${token}`
+      },
       body: JSON.stringify({ status: "completed" }),
     });
     return await res.json();
@@ -46,9 +61,14 @@ export const CompleteAppointment = async (id) => {
 
 export const UpdateAppointment = async (id, data) => {
   try {
+    const token = localStorage.getItem("token"); //  JWT token
+    if (!token) return { success: false, message: "User not logged in" };
     const res = await fetch(`${API_URL}/update/${id}`, {
       method: "PUT",
-      headers: { "Content-Type": "application/json" },
+      headers: {
+        "Content-Type": "application/json",
+        'Authorization': `Bearer ${token}`
+      },
       body: JSON.stringify(data),
     });
     return await res.json();
@@ -60,8 +80,11 @@ export const UpdateAppointment = async (id, data) => {
 
 export const DeleteAppointment = async (id) => {
   try {
+    const token = localStorage.getItem("token");
+    if (!token) return { success: false, message: "User not logged in" };
     const res = await fetch(`${API_URL}/delete/${id}`, {
       method: "DELETE",
+      headers: { "Content-Type": "application/json", 'Authorization': `Bearer ${token}` },
     });
     return await res.json();
   } catch (err) {
@@ -72,15 +95,20 @@ export const DeleteAppointment = async (id) => {
 
 
 export const GetAllAppointmentById = async (id) => {
-    try {
-        const res = await fetch(`${API_URL}/${id}`, {
-            method: "GET",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify(),
-        });
-        return await res.json();
-    } catch (err) {
-        console.error("GetAppointment By Id error:", err);
-        return { success: false, message: "Network error" };
-    }
+  try {
+    const token = localStorage.getItem("token"); //  JWT token
+    if (!token) return { success: false, message: "User not logged in" };
+    const res = await fetch(`${API_URL}/${id}`, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+        'Authorization': `Bearer ${token}`
+      },
+
+    });
+    return await res.json();
+  } catch (err) {
+    console.error("GetAppointment By Id error:", err);
+    return { success: false, message: "Network error" };
+  }
 };
