+++ frontend/src/api/BloodBankReqApi.js
@@ -1,12 +1,17 @@
 const API_URL = "https://hospital-patient-care-dashboard-backend.onrender.com/api/bloodBankReq";
 
 // get all confirmed appointments
-export const GetConfrimedAppointment=async () => {
+export const GetConfrimedAppointment = async () => {
     try {
+        const token = localStorage.getItem("token"); //  JWT token
+        if (!token) return { success: false, message: "User not logged in" };
         const res = await fetch(`${API_URL}/confirmed-appointments`, {
             method: "GET",
-            headers: { "Content-Type": "application/json" },
-           
+            headers: {
+                "Content-Type": "application/json",
+                'Authorization': `Bearer ${token}`
+            },
+
         });
         return await res.json();
     } catch (err) {
@@ -16,33 +21,39 @@ export const GetConfrimedAppointment=async () => {
 };
 
 
-
-
-
 // create req
-export const CreateBloodReq = async (data) =>{
-    try{
- const res = await fetch(`${API_URL}/create`, {
+export const CreateBloodReq = async (data) => {
+    try {
+        const token = localStorage.getItem("token"); //  JWT token
+        if (!token) return { success: false, message: "User not logged in" };
+        const res = await fetch(`${API_URL}/create`, {
             method: "POST",
-            headers: { "Content-Type": "application/json" },
-              body: JSON.stringify(data),
+            headers: {
+                "Content-Type": "application/json",
+                'Authorization': `Bearer ${token}`
+            },
+            body: JSON.stringify(data),
         });
-      const result = await res.json(); 
-    return result;
+        const result = await res.json();
+        return result;
     }
-    catch(err)
-    {
-         console.error("req  error:", err);
+    catch (err) {
+        console.error("req  error:", err);
         return { success: false, message: "Network error" };
     }
 }
 // get all req
 export const GetAllBloodReq = async () => {
     try {
+        const token = localStorage.getItem("token"); //  JWT token
+        if (!token) return { success: false, message: "User not logged in" };
         const res = await fetch(`${API_URL}/`, {
             method: "GET",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify(),
+            headers: {
+                "Content-Type": "application/json",
+                'Authorization': `Bearer ${token}`
+            },
+
         });
         return await res.json();
     } catch (err) {
@@ -52,11 +63,16 @@ export const GetAllBloodReq = async () => {
 };
 
 // updateReq
-export const UpdateBloodReq= async(id,data)=>{
-     try {
+export const UpdateBloodReq = async (id, data) => {
+    try {
+        const token = localStorage.getItem("token"); //  JWT token
+        if (!token) return { success: false, message: "User not logged in" };
         const res = await fetch(`${API_URL}/update/${id}`, {
             method: "PUT",
-            headers: { "Content-Type": "application/json" },
+            headers: {
+                "Content-Type": "application/json",
+                'Authorization': `Bearer ${token}`
+            },
             body: JSON.stringify(data),
         });
         return await res.json();
@@ -68,12 +84,17 @@ export const UpdateBloodReq= async(id,data)=>{
 
 
 // Delete Req
-export const DeleteBloodReq=async(id)=>{
-     try {
+export const DeleteBloodReq = async (id) => {
+    try {
+        const token = localStorage.getItem("token"); //  JWT token
+        if (!token) return { success: false, message: "User not logged in" };
         const res = await fetch(`${API_URL}/delete/${id}`, {
             method: "DELETE",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify(),
+            headers: {
+                "Content-Type": "application/json",
+                'Authorization': `Bearer ${token}`
+            },
+
         });
         return await res.json();
     } catch (err) {
