+++ frontend/src/api/HospStaff.js
@@ -1,14 +1,21 @@
+const API_URL = "https://hospital-patient-care-dashboard-backend.onrender.com/api/hospStaff";
+
+
+
 export const RegisterHospStaff = async ({ email, dept_id, name, phone, medical_license, exp_years, specialization_id }) => {
 
    try {
-      const res = await fetch('https://hospital-patient-care-dashboard-backend.onrender.com/api/hospStaff/register', {
+      const token = localStorage.getItem("token"); //  JWT token
+      if (!token) return { success: false, message: "User not logged in" };
+      const res = await fetch(`${API_URL}/register`, {
          method: 'POST',
-         headers: { 'Content-Type': 'application/json' },
+         headers: { 'Content-Type': 'application/json' ,
+                'Authorization': `Bearer ${token}`},
          body: JSON.stringify({ email, dept_id, name, phone, medical_license, exp_years, specialization_id })
-        
+
 
       })
-       console.log("RegisterHospStaff sending:", { email, dept_id, name, phone, medical_license, exp_years, specialization_id });
+      console.log("RegisterHospStaff sending:", { email, dept_id, name, phone, medical_license, exp_years, specialization_id });
       const result = await res.json()
       return result
    }
@@ -19,29 +26,38 @@ export const RegisterHospStaff = async ({ email, dept_id, name, phone, medical_l
 }
 
 export const GetAllHospStaff = async () => {
-   const res = await fetch('https://hospital-patient-care-dashboard-backend.onrender.com/api/hospStaff/', {
+   const token = localStorage.getItem("token"); //  JWT token
+   if (!token) return { success: false, message: "User not logged in" };
+   const res = await fetch(`${API_URL}/`, {
       method: 'GET',
-      headers: { 'Content-Type': 'application/json' }
-      //  body:JSON.stringify()
+      headers: { 'Content-Type': 'application/json' ,
+                'Authorization': `Bearer ${token}`}
+
    })
    const result = await res.json()
    return result
 }
 
 export const getHospStaffById = async (id) => {
-   const res = await fetch(`https://hospital-patient-care-dashboard-backend.onrender.com/api/hospStaff/${id}`, {
+   const token = localStorage.getItem("token"); //  JWT token
+   if (!token) return { success: false, message: "User not logged in" };
+   const res = await fetch(`${API_URL}/${id}`, {
       method: 'GET',
-      headers: { 'Content-Type': 'application/json' }
-      // body: JSON.stringify()
+      headers: { 'Content-Type': 'application/json' ,
+                'Authorization': `Bearer ${token}`}
+
    })
    const result = await res.json()
    return result
 }
 
 export const deleteHospStaff = async (id) => {
-   const res = await fetch(`https://hospital-patient-care-dashboard-backend.onrender.com/api/hospStaff/delete/${id}`, {
+   const token = localStorage.getItem("token"); //  JWT token
+   if (!token) return { success: false, message: "User not logged in" };
+   const res = await fetch(`${API_URL}/delete/${id}`, {
       method: 'DELETE',
-      headers: { 'Content-Type': 'application/json' }
+      headers: { 'Content-Type': 'application/json' ,
+                'Authorization': `Bearer ${token}`}
 
    })
    const result = await res.json()
@@ -50,9 +66,12 @@ export const deleteHospStaff = async (id) => {
 
 
 export const updateHospStaff = async (id, { dept_id, name, phone, medical_license, exp_years, specialization_id, email }) => {
-   const res = await fetch(`https://hospital-patient-care-dashboard-backend.onrender.com/api/hospStaff/update/${id}`, {
+   const token = localStorage.getItem("token"); //  JWT token
+   if (!token) return { success: false, message: "User not logged in" };
+   const res = await fetch(`${API_URL}/update/${id}`, {
       method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
+      headers: { 'Content-Type': 'application/json' ,
+                'Authorization': `Bearer ${token}`},
       body: JSON.stringify({ dept_id, name, phone, medical_license, exp_years, specialization_id, email })
    })
    const result = await res.json()
