+++ frontend/src/api/SpecializationMasterApi.js
@@ -1,28 +1,41 @@
-export const AddSpecializations= async({spec_name})=>{
- const res=await fetch('https://hospital-patient-care-dashboard-backend.onrender.com/api/specializationMaster/create',{
-    method:'POST',
-    headers:{'Content-Type':'application/json'},
-    body:JSON.stringify({spec_name})
- })
- const result= await res.json()
- return result
+
+const API_URL = "https://hospital-patient-care-dashboard-backend.onrender.com/api/specializationMaster";
+
+
+export const AddSpecializations = async ({ spec_name }) => {
+   const token = localStorage.getItem("token"); //  JWT token
+   if (!token) return { success: false, message: "User not logged in" };
+   const res = await fetch(`${API_URL}/create`, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' ,
+                'Authorization': `Bearer ${token}`},
+      body: JSON.stringify({ spec_name })
+   })
+   const result = await res.json()
+   return result
 }
 
-export const getAllSpecializations= async()=>{
- const res=await fetch('https://hospital-patient-care-dashboard-backend.onrender.com/api/specializationMaster/',{
-   method:'GET',
-    headers:{'Content-Type':'application/json'},
-    body:JSON.stringify()
- })
-  const result= await res.json()
- return result
+export const getAllSpecializations = async () => {
+   const token = localStorage.getItem("token"); //  JWT token
+   if (!token) return { success: false, message: "User not logged in" };
+   const res = await fetch(`${API_URL}/`, {
+      method: 'GET',
+      headers: { 'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}` },
+
+   })
+   const result = await res.json()
+   return result
 }
 
 export const deleteSpecialization = async (id) => {
-   const res = await fetch(`https://hospital-patient-care-dashboard-backend.onrender.com/api/specializationMaster/delete/${id}`, {
+   const token = localStorage.getItem("token"); //  JWT token
+   if (!token) return { success: false, message: "User not logged in" };
+   const res = await fetch(`${API_URL}/delete/${id}`, {
       method: 'DELETE',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify()
+      headers: { 'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}` },
+
    })
    const result = await res.json()
    return result
